% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_LD.R
\name{calculate_LD}
\alias{calculate_LD}
\title{Generates a data frame of pairwise linkage disequilibrium statistics}
\usage{
calculate_LD(dat, genome_build = "grch37")
}
\arguments{
\item{dat}{a dataframe  including SNP column such as:variants while all SNP must be on the same chromosome, using an rsID or chromosome coordinate}

\item{genome_build}{Choose between one of the three options...'grch37' for genome build GRCh37 (hg19), 'grch38' for GRCh38 (hg38), or 'grch38_high_coverage' for GRCh38 High Coverage (hg38) 1000 Genome Project data sets. Default is GRCh37 (hg19).}
}
\value{
a data frame
}
\description{
Generates a data frame of pairwise linkage disequilibrium statistics
}
