% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_plot.R
\name{circle_plot}
\alias{circle_plot}
\title{Return a circle plot}
\usage{
circle_plot(a, padj = FALSE)
}
\arguments{
\item{a}{The combine_results()  function returned a result}

\item{padj}{the padj plot}
}
\value{
Return a circle plot
}
\description{
Return a circle plot
}
\examples{
# p+expand_limits(x = c(-100, 400), y = c(0, 4))
}
