% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_local.R
\name{coloc_local}
\alias{coloc_local}
\title{Prepare datasets to Bayesian colocalisation analysis using Bayes Factors}
\usage{
coloc_local(
  eqtl,
  gwas,
  beta_col_eqtl = "beta",
  se_col_eqtl = "varbeta",
  snp_col_eqtl = "snp",
  pval_col_eqtl = "pval",
  MAF_col_eqtl = "MAF",
  N_eqtl = NULL,
  type_eqtl = NULL,
  s_eqtl = NULL,
  sdY_eqtl = NULL,
  beta_col_gwas = "beta",
  se_col_gwas = "varbeta",
  snp_col_gwas = "snp",
  pval_col_gwas = "pval",
  MAF_col_gwas = "MAF",
  N_gwas = NULL,
  type_gwas = NULL,
  s_gwas = NULL,
  sdY_gwas = NULL
)
}
\arguments{
\item{eqtl}{the dataframe of eqtl}

\item{gwas}{the dataframe of gwas}

\item{beta_col_eqtl}{regression coefficient for each SNP from dataset1}

\item{se_col_eqtl}{variance of beta from dataset1}

\item{snp_col_eqtl}{a character vector of snp ids, optional. It will be used to merge dataset1 and dataset2 and will be retained in the results.}

\item{pval_col_eqtl}{P-values for each SNP in dataset1}

\item{MAF_col_eqtl}{minor allele frequency of the variants}

\item{N_eqtl}{Number of samples in dataset 1}

\item{type_eqtl}{the type of data in dataset 1 - either "quant" or "cc" to denote quantitative or case-control}

\item{s_eqtl}{for a case control dataset, the proportion of samples in dataset 1 that are cases}

\item{sdY_eqtl}{for a quantitative trait, the population standard deviation of the trait. if not given, it can be estimated from the vectors of varbeta and MAF}

\item{beta_col_gwas}{regression coefficient for each SNP from gwas}

\item{se_col_gwas}{variance of beta from gwas}

\item{snp_col_gwas}{a character vector of snp ids, optional. It will be used to merge dataset1 and dataset2 and will be retained in the results.}

\item{pval_col_gwas}{P-values for each SNP in dataset1}

\item{MAF_col_gwas}{minor allele frequency of the variants}

\item{N_gwas}{Number of samples in dataset 2}

\item{type_gwas}{the type of data in dataset 2 - either "quant" or "cc" to denote quantitative or case-control}

\item{s_gwas}{for a case control dataset, the proportion of samples in dataset 2 that are cases}

\item{sdY_gwas}{for a quantitative trait, the population standard deviation of the trait. if not given, it can be estimated from the vectors of varbeta and MAF}
}
\value{
the list of dataset 1 and 2
}
\description{
Prepare datasets to Bayesian colocalisation analysis using Bayes Factors
}
\examples{
# also can:
# check_dataset(result$dataset1)
# check_dataset(result$dataset2)
# my.res <- coloc.abf(result$dataset1, result$dataset2)

# only using the pval and SNP and maf
# result_coloc <- coloc.abf(dataset1=list(snp=result$dataset1$snp,pvalues=result$dataset1$pvalues,type="cc", s=0.001, N=359399), dataset2=list(snp=result$dataset2$snp,pvalues=result$dataset2$pvalues, type="quant", N=31684), MAF=result$dataset1$MAF)
}
