% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{combine_results}
\alias{combine_results}
\title{Combination the result,test,presso results}
\usage{
combine_results(result = NULL, test = NULL, presso = NULL)
}
\arguments{
\item{result}{The IVW_fix_random() function returned a result}

\item{test}{The mr_test()  function returned a result}

\item{presso}{The MRpresso()  function returned a result}
}
\value{
Combination the result,test,presso results
}
\description{
Combination the result,test,presso results
}
