% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_finngen_exposure.R
\name{finngen_exposure_data}
\alias{finngen_exposure_data}
\title{Return exposure when using the finngen database}
\usage{
finngen_exposure_data(trait, filepath, samplesize, p_threshold = 5e-08)
}
\arguments{
\item{trait}{the name of exposure, eg: osteonecrosis}

\item{filepath}{the exposure original file path}

\item{samplesize}{the exposure samplesize}

\item{p_threshold}{the p_threshold = 5e-8}
}
\value{
a dataframe of finngen exposure IVs
}
\description{
Return exposure when using the finngen database
}
