% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_f.R
\name{get_f}
\alias{get_f}
\title{Return the IVs (F>10)}
\usage{
get_f(dat, F_value = 10)
}
\arguments{
\item{dat}{the exposure or harmonization between exposure and outcome}

\item{F_value}{the F statistic = 10}
}
\value{
Calculate the F statistic for the exposure IVs (F>10)
}
\description{
Return the IVs (F>10)
}
