% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rsID_from_chrpos.R
\name{get_rsID_from_chrpos}
\alias{get_rsID_from_chrpos}
\title{How to get rsID from chromosome and position}
\usage{
get_rsID_from_chrpos(
  dat,
  col_chr = "chr",
  col_pos = "start",
  col_ref_allele = "refAllele",
  col_alt_allele = "altAllele",
  col_snp = "SNP",
  build = "37",
  pop = "EUR",
  database = "dbsnp"
)
}
\arguments{
\item{dat}{the dataframe of exposure or outcome}

\item{col_chr}{chromosome the default is "chr_col".}

\item{col_pos}{chromosome position}

\item{col_ref_allele}{Name of column with non effect allele. Must contain only the characters "A", "C", "T" or "G". The default is "refAllele".}

\item{col_alt_allele}{Name of column with effect allele. Must contain only the characters "A", "C", "T" or "G". The default is "altAllele"}

\item{col_snp}{Required name of column with SNP rs IDs. The default is "SNP".}

\item{build}{the version}

\item{pop}{investigated population}

\item{database}{dbsnp database}
}
\value{
a dataframe
}
\description{
How to get rsID from chromosome and position
}
