% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS.R, R/GRS_drug.R
\name{GRS}
\alias{GRS}
\title{Return the genetic risk scores}
\usage{
GRS(dat)

GRS(dat)
}
\arguments{
\item{dat}{the harmonization data}
}
\value{
GRS dataframe

GRS dataframe of drugs Mendelian
}
\description{
Return the genetic risk scores

Return the genetic risk scores for drugs Mendelian
}
