% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRS_drug.R
\name{GRS_drug}
\alias{GRS_drug}
\title{Return the genetic risk scores for drugs Mendelian}
\usage{
GRS_drug(dat)
}
\arguments{
\item{dat}{the harmonization data}
}
\value{
GRS dataframe of drugs Mendelian
}
\description{
Return the genetic risk scores for drugs Mendelian
}
