% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSMR.R
\name{GSMR}
\alias{GSMR}
\title{Return the GSMR analysis}
\usage{
GSMR(dat, plot = FALSE, folder = NULL, LDmatrix = FALSE)
}
\arguments{
\item{dat}{the harmonization data}

\item{plot}{the choice of showing the picture}

\item{folder}{the folder names}

\item{LDmatrix}{The choice of LD correlation matrix}
}
\value{
Return the GSMR analysis dataframe
}
\description{
Return the GSMR analysis
}
\examples{
# https://yanglab.westlake.edu.cn/software/gsmr/
}
