% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVW_fix_random.R
\name{IVW_fix_random}
\alias{IVW_fix_random}
\title{Return the main four MR results}
\usage{
IVW_fix_random(
  dat,
  other_method = c("mr_egger_regression", "mr_weighted_median", "mr_raps"),
  plot = FALSE,
  Random_only = FALSE
)
}
\arguments{
\item{dat}{harmonization(using the function of harmonise_data)}

\item{other_method}{the mr methods:"mr_egger_regression","mr_weighted_median", "mr_raps"}

\item{plot}{the choice of showing the MR results visualization}

\item{Random_only}{the choices of the only use IVW_mre}
}
\description{
Return the main four MR results
}
