% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_test.R
\name{mr_test}
\alias{mr_test}
\title{Return the  Cochran’s Q heterogeneity test and directional pleiotropy test}
\usage{
mr_test(dat, full = TRUE)
}
\arguments{
\item{dat}{harmonization(using the function of harmonise_data) note:the SNP>=3}

\item{full}{the more details of test results or none}
}
\value{
the sensitivity analysis
}
\description{
Return the  Cochran’s Q heterogeneity test and directional pleiotropy test
}
