% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRpresso.R
\name{MRpresso}
\alias{MRpresso}
\title{Return the MR-Presso results of dat (after harmonization)}
\usage{
MRpresso(dat)
}
\arguments{
\item{dat}{harmonization(using the function of harmonise_data)}
}
\value{
the value of MR-Presso results
(exposure,outcome,Presso.Pvalue,Presso.RSSobs,b.MR.presso,se.MR.presso,pval.MR.presso)
}
\description{
Return the MR-Presso results of dat (after harmonization)
}
