% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi}
\alias{multi}
\title{Returns the sum of two numbers}
\usage{
multi(a, b)
}
\arguments{
\item{a}{Number one.}

\item{b}{Number two.}
}
\value{
Sum of two number.
}
\description{
Returns the sum of two numbers
}
\examples{
library(MRZhou)
multi(1, 2)
}
