% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_table.R
\name{paper_table}
\alias{paper_table}
\title{Return the main MR analysis results including:TableS1-3}
\usage{
paper_table(exp, out, dat, result, test, presso)
}
\arguments{
\item{exp}{Exposure instrumental variable}

\item{out}{Outcome}

\item{dat}{harmonization(using the function of harmonise_data)}

\item{result}{The IVW_fix_random() function returned a result}

\item{test}{The mr_test()  function returned a result}

\item{presso}{The MRpresso()  function returned a result}
}
\value{
a paper needing marital
}
\description{
Return the main MR analysis results including:TableS1-3
}
