% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_finngen_outcome.R
\name{read_finngen_outcome}
\alias{read_finngen_outcome}
\title{Return outcome when using the finngen database}
\usage{
read_finngen_outcome(exp, trait, filepath, samplesize)
}
\arguments{
\item{exp}{the exposure IVs selection}

\item{trait}{the name of outcome,eg:osteonecrosis}

\item{filepath}{the outcome original file path}

\item{samplesize}{the outcome samplesize}
}
\value{
a dataframe of finngen outcome
}
\description{
Return outcome when using the finngen database
}
