% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read IEU GWAS database vcf files (original)}
\usage{
read_vcf(path)
}
\arguments{
\item{path}{The path of vcf files}
}
\value{
a dataframe
}
\description{
Read IEU GWAS database vcf files (original)
}
\examples{
# library(gwasvcf)
# library(gwasglue)
# library(VariantAnnotation)
# library(TwoSampleMR)
}
