% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMR_analysis.R
\name{SMR_analysis}
\alias{SMR_analysis}
\title{Summary-data-based Mendelian Randomization(SMR) analysis}
\usage{
SMR_analysis(
  smr_exe_dir = "/public/home/fanfangzhou/apps/SMR/smr_linux",
  g1000_dir = "/public/home/fanfangzhou/R/1kgv3/g1000/EUR",
  qtls_dir,
  input_dir = "gene.list",
  gwas_dir,
  outcome_prefix = "targets_GWAS",
  multi_smr = FALSE,
  linux = TRUE
)
}
\arguments{
\item{smr_exe_dir}{the SMR path}

\item{g1000_dir}{the g1000/EUR path}

\item{qtls_dir}{the eqtl database,eg:eqtlgen,getxv8}

\item{input_dir}{the eqtl gene,eg:eqltgen:ENSG00000113161;GTEx:HMGCR}

\item{gwas_dir}{the vector of GWAS}

\item{outcome_prefix}{the file of results}

\item{multi_smr}{whether to analysis multi-smr}

\item{linux}{the performer choise}
}
\value{
the smr results
}
\description{
Summary-data-based Mendelian Randomization(SMR) analysis
}
\examples{
# SMR_analysis(qtls_dir='/public/home/fanfangzhou/R/GTEx8/SMR/eQTLGen/cis-eQTLs-full_eQTLGen_AF_incl_nr_formatted_20191212.new.txt_besd-dense',gwas_dir=c('LDL.ma','RCS_UKB.ma','RCS_R10.ma'))
}
