% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_data.R
\name{smr_data}
\alias{smr_data}
\title{SMR data prepration}
\usage{
smr_data(
  gwas,
  snp_col = "SNP",
  A1_col = "effect_allele",
  A2_col = "other_allele",
  freq_col = "eaf",
  b_col = "beta",
  se_col = "se",
  p_col = "pval",
  N_col = "samplesize",
  filename
)
}
\arguments{
\item{gwas}{the read_vcf results dataframe}

\item{snp_col}{SNP}

\item{A1_col}{effect_allele}

\item{A2_col}{other_allele}

\item{freq_col}{eaf}

\item{b_col}{beta}

\item{se_col}{se}

\item{p_col}{pval}

\item{N_col}{samplesize}

\item{filename}{filename,eg:'LUAD.ma'}
}
\description{
SMR data prepration
}
