% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_plot.R
\name{smr_plot}
\alias{smr_plot}
\title{SMR plot}
\usage{
smr_plot(
  smr_exe_dir = "/public/home/fanfangzhou/apps/SMR/smr_linux",
  g1000_dir = "/public/home/fanfangzhou/R/1kgv3/g1000/EUR",
  qtls_dir,
  gwas_dir,
  queried_probe,
  glist_dir = "/public/home/fanfangzhou/R/GTEx8/SMR/glist-hg19",
  outcome_prefix,
  gwas_label,
  LocusPlot = F,
  linux = TRUE,
  Rscript = "/public/home/fanfangzhou/R/GTEx8/SMR/plot/plot/plot_SMR.r"
)
}
\arguments{
\item{smr_exe_dir}{the SMR path}

\item{g1000_dir}{the g1000/EUR path}

\item{qtls_dir}{the eqtl database,eg:eqtlgen,getxv8}

\item{gwas_dir}{the vector of GWAS}

\item{queried_probe}{eg:ENSG00000113161}

\item{glist_dir}{eg:files of glist-hg19}

\item{outcome_prefix}{the file of results}

\item{gwas_label}{the plot title label}

\item{LocusPlot}{the choice of plot type}

\item{linux}{the performer choise}

\item{Rscript}{the Rcode of SMR plot}
}
\value{
a plot pdf
}
\description{
SMR plot
}
