% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_add_eaf.R
\name{snp_add_eaf}
\alias{snp_add_eaf}
\title{How to find allele frequency, proxy of a given SNP}
\usage{
snp_add_eaf(dat, build = "37", pop = "EUR")
}
\arguments{
\item{dat}{the exposure dataframe}

\item{build}{the version,"37","38"}

\item{pop}{EUR}
}
\value{
a dataframe
}
\description{
How to find allele frequency, proxy of a given SNP
}
