% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_phenoscanner.R
\name{snp_phenoscanner}
\alias{snp_phenoscanner}
\title{How to search pleiotropy for a given SNP}
\usage{
snp_phenoscanner(dat = exposure_MAP)
}
\arguments{
\item{dat}{the exposure dataframe}
}
\value{
a dataframe of  the pleiotropy for a given SNP
}
\description{
How to search pleiotropy for a given SNP
}
