% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Return a volcano plot}
\usage{
volcano_plot(mydata, adjust = "BH")
}
\arguments{
\item{mydata}{The combine_results()  function returned a result}

\item{adjust}{The choose of 'BH','FDR','None'}
}
\value{
Return a volcano plot
}
\description{
Return a volcano plot
}
\examples{
library(ggrepel)
library(gt)
}
