% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Return the forest plot}
\usage{
forest_plot(data, xlim = c(0, 4), ticks_at = c(0, 0.5, 1, 2, 3))
}
\arguments{
\item{data}{harmonization(using the function of harmonise_data)}

\item{xlim}{the forestploter packages param,xlim=c(0,4)}

\item{ticks_at}{the forestploter packages param, ticks_at=c(0,0.5,1, 2, 3)}
}
\value{
plot
}
\description{
Return the forest plot
}
\examples{
library(grid)
library(readr)
library(forestploter)
}
