% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVW_fix_random.R
\name{IVW_fix_random}
\alias{IVW_fix_random}
\title{Return the main four MR results}
\usage{
IVW_fix_random(dat, plot = FALSE)
}
\arguments{
\item{dat}{harmonization(using the function of harmonise_data)}

\item{plot}{the choice of showing the MR results visualization}
}
\description{
Return the main four MR results
}
